RecordDialog: Replaced pykdeuic4 generated code with a cleaned up ui file. Moved RecordDialog class from dialogs.py to recorddialog.py. Class RecordSettings merged into RecordDialog. RecordDialog is now derived from QDialog instead of KDialog.
