+++ lib/autokey/qtui/recorddialog.py
@@ -1,56 +1,51 @@
-#!/usr/bin/env python
-# coding=UTF-8
-#
-# Generated by pykdeuic4 from recorddialog.ui on Sun Mar  4 11:39:40 2012
-#
-# WARNING! All changes to this file will be lost.
-from PyKDE4 import kdecore
-from PyKDE4 import kdeui
-from PyQt4 import QtCore, QtGui
+# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2018 Thomas Hess <thomas.hess@udo.edu>
 
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    _fromUtf8 = lambda s: s
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 
-class Ui_Form(object):
-    def setupUi(self, Form):
-        Form.setObjectName(_fromUtf8("Form"))
-        Form.resize(400, 300)
-        self.verticalLayout = QtGui.QVBoxLayout(Form)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.label = QtGui.QLabel(Form)
-        self.label.setObjectName(_fromUtf8("label"))
-        self.verticalLayout.addWidget(self.label)
-        self.recKeyboardButton = QtGui.QCheckBox(Form)
-        self.recKeyboardButton.setObjectName(_fromUtf8("recKeyboardButton"))
-        self.verticalLayout.addWidget(self.recKeyboardButton)
-        self.recMouseButton = QtGui.QCheckBox(Form)
-        self.recMouseButton.setObjectName(_fromUtf8("recMouseButton"))
-        self.verticalLayout.addWidget(self.recMouseButton)
-        self.horizontalLayout = QtGui.QHBoxLayout()
-        self.horizontalLayout.setObjectName(_fromUtf8("horizontalLayout"))
-        self.label_2 = QtGui.QLabel(Form)
-        self.label_2.setObjectName(_fromUtf8("label_2"))
-        self.horizontalLayout.addWidget(self.label_2)
-        self.secondsSpinBox = KIntSpinBox(Form)
-        self.secondsSpinBox.setProperty("value", 5)
-        self.secondsSpinBox.setObjectName(_fromUtf8("secondsSpinBox"))
-        self.horizontalLayout.addWidget(self.secondsSpinBox)
-        self.label_3 = QtGui.QLabel(Form)
-        self.label_3.setObjectName(_fromUtf8("label_3"))
-        self.horizontalLayout.addWidget(self.label_3)
-        self.verticalLayout.addLayout(self.horizontalLayout)
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
-        self.retranslateUi(Form)
-        QtCore.QMetaObject.connectSlotsByName(Form)
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-    def retranslateUi(self, Form):
-        Form.setWindowTitle(kdecore.i18n(_fromUtf8("Form")))
-        self.label.setText(kdecore.i18n(_fromUtf8("Record a keyboard/mouse macro")))
-        self.recKeyboardButton.setText(kdecore.i18n(_fromUtf8("Record keyboard events")))
-        self.recMouseButton.setText(kdecore.i18n(_fromUtf8("Record mouse events (experimental)")))
-        self.label_2.setText(kdecore.i18n(_fromUtf8("Start recording after")))
-        self.label_3.setText(kdecore.i18n(_fromUtf8("seconds")))
+from .common import inherits_from_ui_file_with_name
+import logging
+logger = logging.getLogger("root").getChild("Qt-GUI").getChild("Record Dialog")  # type: logging.Logger
 
-from PyKDE4.kdeui import KIntSpinBox
+
+class RecordDialog(*inherits_from_ui_file_with_name("record_dialog")):
+
+    def __init__(self, parent, closure):
+        super().__init__(parent)
+        self.setupUi(self)
+        self.closure = closure
+
+    def get_record_keyboard(self):
+        return self.recKeyboardButton.isChecked()
+
+    def get_record_mouse(self):
+        return self.recMouseButton.isChecked()
+
+    def get_delay(self):
+        return self.secondsSpinBox.value()
+
+    def accept(self):
+        super().accept()
+        logger.info("Dialog accepted: Record keyboard: {}, record mouse: {}, delay: {} s".format(
+            self.get_record_keyboard(), self.get_record_mouse(), self.get_delay()
+        ))
+        self.closure(True, self.get_record_keyboard(), self.get_record_mouse(), self.get_delay())
+
+    def reject(self):
+        super().reject()
+        logger.info("Dialog closed (rejected/aborted): Record keyboard: {}, record mouse: {}, delay: {} s".format(
+            self.get_record_keyboard(), self.get_record_mouse(), self.get_delay()
+        ))
+        self.closure(False, self.get_record_keyboard(), self.get_record_mouse(), self.get_delay())
+
